requirements.txt
Appium-Python-Client==5.2.5
attrs==25.4.0
certifi==2026.1.4
exceptiongroup==1.3.1
h11==0.16.0
idna==3.11
iniconfig==2.1.0
Jinja2==3.1.6
MarkupSafe==3.0.3
outcome==1.3.0.post0
packaging==26.0
pluggy==1.6.0
Pygments==2.19.2
PySocks==1.7.1
pytest==8.4.2
pytest-html==4.2.0
pytest-metadata==3.1.1
selenium==4.36.0
sniffio==1.3.1
sortedcontainers==2.4.0
tomli==2.4.0
trio==0.31.0
trio-websocket==0.12.2
typing_extensions==4.15.0
urllib3==2.6.3
websocket-client==1.9.0
wsproto==1.2.0
